Rename ParserHelpers into JavascriptParserHelpers
