+++ lib/APIPlugin.js
@@ -9,7 +9,7 @@ const {
 	toConstantDependency,
 	toConstantDependencyWithWebpackRequire,
 	evaluateToString
-} = require("./ParserHelpers");
+} = require("./JavascriptParserHelpers");
 
 const NullFactory = require("./NullFactory");
