+++ lib/ConstPlugin.js
@@ -6,7 +6,7 @@
 
 const ConstDependency = require("./dependencies/ConstDependency");
 const NullFactory = require("./NullFactory");
-const { evaluateToString } = require("./ParserHelpers");
+const { evaluateToString } = require("./JavascriptParserHelpers");
 
 const getQuery = request => {
 	const i = request.indexOf("?");
