+++ lib/DefinePlugin.js
@@ -11,7 +11,7 @@ const {
 	evaluateToString,
 	toConstantDependency,
 	toConstantDependencyWithWebpackRequire
-} = require("./ParserHelpers");
+} = require("./JavascriptParserHelpers");
 const NullFactory = require("./NullFactory");
 
 const stringifyObj = obj => {
