+++ lib/ExtendedAPIPlugin.js
@@ -9,7 +9,7 @@ const ConstDependency = require("./dependencies/ConstDependency");
 const {
 	evaluateToString,
 	toConstantDependencyWithWebpackRequire
-} = require("./ParserHelpers");
+} = require("./JavascriptParserHelpers");
 const NullFactory = require("./NullFactory");
 
 const REPLACEMENTS = {
