+++ lib/HotModuleReplacementPlugin.js
@@ -16,7 +16,7 @@ const {
 	evaluateToString,
 	skipTraversal,
 	toConstantDependencyWithWebpackRequire
-} = require("./ParserHelpers");
+} = require("./JavascriptParserHelpers");
 
 module.exports = class HotModuleReplacementPlugin {
 	constructor(options) {
