+++ lib/JavascriptParserHelpers.js
