+++ lib/NodeStuffPlugin.js
@@ -13,7 +13,7 @@ const {
 	expressionIsUnsupported,
 	toConstantDependency,
 	toConstantDependencyWithWebpackRequire
-} = require("./ParserHelpers");
+} = require("./JavascriptParserHelpers");
 const NullFactory = require("./NullFactory");
 
 class NodeStuffPlugin {
