+++ lib/ProvidePlugin.js
@@ -9,7 +9,7 @@ const {
 	addParsedVariableToModule,
 	approve,
 	toConstantDependency
-} = require("./ParserHelpers");
+} = require("./JavascriptParserHelpers");
 const NullFactory = require("./NullFactory");
 
 class ProvidePlugin {
