+++ lib/RequireJsStuffPlugin.js
@@ -9,7 +9,7 @@ const NullFactory = require("./NullFactory");
 const {
 	toConstantDependency,
 	toConstantDependencyWithWebpackRequire
-} = require("./ParserHelpers");
+} = require("./JavascriptParserHelpers");
 
 module.exports = class RequireJsStuffPlugin {
 	apply(compiler) {
