+++ lib/dependencies/AMDPlugin.js
@@ -25,7 +25,7 @@ const {
 	evaluateToIdentifier,
 	evaluateToString,
 	toConstantDependency
-} = require("../ParserHelpers");
+} = require("../JavascriptParserHelpers");
 
 class AMDPlugin {
 	constructor(options, amdOptions) {
