+++ lib/dependencies/CommonJsPlugin.js
@@ -20,7 +20,7 @@ const {
 	evaluateToIdentifier,
 	evaluateToString,
 	toConstantDependency
-} = require("../ParserHelpers");
+} = require("../JavascriptParserHelpers");
 
 class CommonJsPlugin {
 	constructor(options) {
