+++ lib/dependencies/CommonJsRequireDependencyParserPlugin.js
@@ -10,7 +10,7 @@ const RequireHeaderDependency = require("./RequireHeaderDependency");
 const LocalModuleDependency = require("./LocalModuleDependency");
 const ContextDependencyHelpers = require("./ContextDependencyHelpers");
 const LocalModulesHelpers = require("./LocalModulesHelpers");
-const { toConstantDependencyWithWebpackRequire } = require("../ParserHelpers");
+const { toConstantDependencyWithWebpackRequire } = require("../JavascriptParserHelpers");
 
 class CommonJsRequireDependencyParserPlugin {
 	constructor(options) {
