+++ lib/dependencies/RequireEnsurePlugin.js
@@ -11,7 +11,7 @@ const NullFactory = require("../NullFactory");
 
 const RequireEnsureDependenciesBlockParserPlugin = require("./RequireEnsureDependenciesBlockParserPlugin");
 
-const { evaluateToString, toConstantDependency } = require("../ParserHelpers");
+const { evaluateToString, toConstantDependency } = require("../JavascriptParserHelpers");
 
 class RequireEnsurePlugin {
 	apply(compiler) {
