+++ lib/dependencies/RequireIncludePlugin.js
@@ -7,7 +7,7 @@
 const RequireIncludeDependency = require("./RequireIncludeDependency");
 const RequireIncludeDependencyParserPlugin = require("./RequireIncludeDependencyParserPlugin");
 
-const { evaluateToString, toConstantDependency } = require("../ParserHelpers");
+const { evaluateToString, toConstantDependency } = require("../JavascriptParserHelpers");
 
 class RequireIncludePlugin {
 	apply(compiler) {
