+++ lib/dependencies/SystemPlugin.js
@@ -10,7 +10,7 @@ const {
 	expressionIsUnsupported,
 	requireFileAsExpression,
 	toConstantDependency
-} = require("../ParserHelpers");
+} = require("../JavascriptParserHelpers");
 const WebpackError = require("../WebpackError");
 
 class SystemPlugin {
