+++ lib/node/NodeSourcePlugin.js
@@ -9,7 +9,7 @@ const nodeLibsBrowser = require("node-libs-browser");
 const {
 	addParsedVariableToModule,
 	requireFileAsExpression
-} = require("../ParserHelpers");
+} = require("../JavascriptParserHelpers");
 
 module.exports = class NodeSourcePlugin {
 	constructor(options) {
